+++ nodcms-core/Config/Boot/development.php
@@ -39,13 +39,4 @@ defined('SHOW_DEBUG_BACKTRACE') || define('SHOW_DEBUG_BACKTRACE', true);
   | the system. This will control whether Kint is loaded, and a few other
   | items. It can always be used within your own application too.
  */
-
 defined('CI_DEBUG') || define('CI_DEBUG', true);
-
-/*
- * NodCMS accept test payment
- *
- * This define create a test payment method to test payment routines
- */
-
-defined('ALLOW_TEST_PAYMENT') || define("ALLOW_TEST_PAYMENT", true);