+++ public/index.php
@@ -10,43 +10,11 @@
  *
  */
 
-// Valid PHP Version?
-$minPHPVersion = '7.3';
-if (phpversion() < $minPHPVersion)
-{
-	die("Your PHP version must be {$minPHPVersion} or higher to run CodeIgniter. Current version: " . phpversion());
-}
-unset($minPHPVersion);
-
 // Path to the front controller (this file)
 define('FCPATH', __DIR__ . DIRECTORY_SEPARATOR);
 
-// Root directory
-define("ROOTPATH", dirname(FCPATH).DIRECTORY_SEPARATOR);
-// CodeIgniter core path
-define("SYSTEMPATH", ROOTPATH."vendor/codeigniter4/framework/system".DIRECTORY_SEPARATOR);
-// NodCMS core path
-define("COREPATH", ROOTPATH."nodcms-core".DIRECTORY_SEPARATOR);
-// NodCMS public path
-define("SELF_PATH", FCPATH);
-
-// Find the requested protocol
-$protocol_status = intval(isset($_SERVER['HTTPS']));
-define("SSL_PROTOCOL", $protocol_status);
-define("URL_PROTOCOL", $protocol_status ? "https://" : "http://");
-
-// Find the base url
-define("CONFIG_BASE_URL", URL_PROTOCOL.$_SERVER['HTTP_HOST'] . rtrim(dirname(str_replace('index.php/', '', $_SERVER['PHP_SELF'])), '/\\') . "/");
-
-define("DB_CONFIG_PATH", COREPATH.'Config/Database.php');
-
-// Location of the NodCMS addon bootstrap file.
-require COREPATH.'bootstrap.php';
-
-// Location of the Paths config file.
-// This is the line that might need to be changed, depending on your folder structure.
-$pathsPath = realpath(COREPATH . 'Config/Paths.php');
-// ^^^ Change this if you move your application folder
+// Ensure the current directory is pointing to the front controller's directory
+chdir(FCPATH);
 
 /*
  *---------------------------------------------------------------
@@ -57,20 +25,21 @@ $pathsPath = realpath(COREPATH . 'Config/Paths.php');
  * and fires up an environment-specific bootstrapping.
  */
 
-// Ensure the current directory is pointing to the front controller's directory
-chdir(__DIR__);
-
 // Load our paths config file
-require $pathsPath;
+// This is the line that might need to be changed, depending on your folder structure.
+require FCPATH . '../nodcms-core/Config/Paths.php';
+// ^^^ Change this line if you move your application folder
+
 $paths = new Config\Paths();
 
-// Location of the framework bootstrap file.
-require rtrim($paths->systemDirectory, '\\/ ') . DIRECTORY_SEPARATOR . 'bootstrap.php';
-
+// Location of the NodCMS bootstrap file.
+require rtrim($paths->appDirectory, '\\/ ') . DIRECTORY_SEPARATOR . 'bootstrap.php';
 
 // Load environment settings from .env files into $_SERVER and $_ENV
 require_once SYSTEMPATH . 'Config/DotEnv.php';
 (new CodeIgniter\Config\DotEnv(ROOTPATH))->load();
+
+define('DB_CONFIG_PATH', '../'.getenv('app.dbConfigPath'));
 
 /*
  * ---------------------------------------------------------------